+++ src/Structure/SurfaceMesh/Membrane.cpp
@@ -292,7 +292,7 @@ double Membrane::signedDistance(const std::array<double, 3>& p, bool safe)const 
     }
     else {
         // Find the current compartment indices containing point p
-        vector<int> indices = GController::getCompartmentIndices(mathfunc::array2Vector(p));
+        vector<size_t> indices = GController::getCompartmentIndices(mathfunc::array2Vector(p));
 
         if(indices.size() != 3)
             throw std::logic_error("Only 3D compartments are allowed in the membrane signed distance calculation.");
@@ -300,15 +300,15 @@ double Membrane::signedDistance(const std::array<double, 3>& p, bool safe)const 
 		unordered_set<Triangle*> triSet; // Set of triangles to be searched
 
         // Find all the neighbor compartments and find triangles to be added to search set.
-        // The for loops are written like this for aesthetic reasons. Be very careful with the scope below.
+        // The for loops are written like this for aesthetic reasons. Be careful with the scope below.
         for(int v0 = -1; v0 <= 1; ++v0) for(int v1 = -1; v1 <= 1; ++v1) for(int v2 = -1; v2 <= 1; ++v2) {
 
-            vector<int> newIndices = {indices[0] + v0, indices[1] + v1, indices[2] + v2};
+            vector<size_t> newIndices = {indices[0] + v0, indices[1] + v1, indices[2] + v2};
             if(GController::indicesOutOfBound(indices)) {
                 // Compartment not found. Simply ignore it.
                 continue;
             }
-            
+
             Compartment* c = GController::getCompartment(newIndices);
 
             // Compartment exists
